Fixed MainContainer and reorganized files
