+++ navigation/MainContainer.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { NavigationContainer } from '@react-navigation/native';
+import { NavigationContainer, DefaultTheme } from '@react-navigation/native';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import Icon from 'react-native-ionicons';
 
@@ -10,40 +10,42 @@ import Details from './screens/Details';
 import ProfileScreen from './screens/ProfileScreen';
 import SettingsScreen from './screens/SettingsScreen';
 
-// Screen Names
-const homeName = 'Home';
-const detailsName = 'Details';
-const profileName = 'Profile';
-const settingsScreen = 'Settings';
-
 const Tab = createBottomTabNavigator();
 
+const theme = {
+    ...DefaultTheme,
+    colors: {
+      ...DefaultTheme.colors,
+      background: "transparent"
+    }
+  }
 
 const MainContainer = () => {
     return (
-      <NavigationContainer theme={theme}>
+        <NavigationContainer theme={theme}>
         <Tab.Navigator
-        initialRouteName={homeName}
-        screenOptions={({route}) => ({
-            tabBarIcon: ({focused, color, size}) => {
-                let iconName;
-                let rn = route.name;
+        initialRouteName={'Home'}
+        // screenOptions={({route}) => ({
+        //     tabBarIcon: ({focused, color, size}) => {
+        //         let iconName;
+        //         let rn = route.name;
 
-                if (rn == homeName) {
-                    iconName = focused ? 'home' : 'home-outline'
-                } else if (rn == detailsName) {
-                    iconName = focused ? 'list' : 'list-outline'
-                } else if (rn == settingsScreen) {
-                    iconName = focused ? 'settings' : 'settings-outline'
-                }
+        //         if (rn == 'Home') {
+        //             iconName = focused ? 'home' : 'home-outline'
+        //         } else if (rn == 'Details') {
+        //             iconName = focused ? 'list' : 'list-outline'
+        //         } else if (rn == 'Settings') {
+        //             iconName = focused ? 'settings' : 'settings-outline'
+        //         }
 
-                return <Icon name={iconName} size={size} color={color}/>
-            },
-        })}>
-            <Tab.Screen name={homeName} component={Home}/>
-            <Tab.Screen name={detailsName} component={Details}/>
-            <Tab.Screen name={profileName} component={ProfileScreen}/>
-            <Tab.Screen name={settingsScreen} component={SettingsScreen}/>
+        //         return <Icon name={iconName} size={size} color={color}/>
+        //     },
+        // })}
+        >
+            <Tab.Screen name={'Home'} component={Home}/>
+            <Tab.Screen name={'Details'} component={Details}/>
+            <Tab.Screen name={'Profile'} component={ProfileScreen}/>
+            <Tab.Screen name={'Settings'} component={SettingsScreen}/>
 
         </Tab.Navigator>
       </NavigationContainer>
